+++ test/fixtures/riverpod_shop_app/lib/repositories/catalog_repository.dart
@@ -1,0 +1,9 @@
+import '../services/catalog_service.dart';
+
+class CatalogRepository {
+  CatalogRepository(this._service);
+
+  final CatalogService _service;
+
+  Future<List<String>> loadCatalog() => _service.fetchItems();
+}